Move blipSyncContext into replicator package
