+++ replicator/blip_sync_context_test.go
@@ -1,4 +1,4 @@
-package rest
+package replicator
 
 import (
 	"context"
@@ -37,7 +37,7 @@ func TestBlipSyncContextSetUseDeltas(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			ctx := &blipSyncContext{
+			ctx := &BlipSyncContext{
 				dbStats:        db.NewDatabaseStats(),
 				blipContextDb:  &db.Database{Ctx: context.TODO()},
 				useDeltas:      tt.startingCtxDeltas,
@@ -78,7 +78,7 @@ func BenchmarkBlipSyncContextSetUseDeltas(b *testing.B) {
 
 	for _, tt := range tests {
 		b.Run(tt.name, func(b *testing.B) {
-			ctx := &blipSyncContext{
+			ctx := &BlipSyncContext{
 				dbStats:       db.NewDatabaseStats(),
 				blipContextDb: &db.Database{Ctx: context.TODO()},
 			}
