+++ rest/blip_api_test.go
@@ -12,6 +12,7 @@ import (
 	"testing"
 	"time"
 
+	"github.com/couchbase/sync_gateway/replicator"
 	"github.com/stretchr/testify/require"
 
 	"github.com/couchbase/go-blip"
@@ -548,7 +549,7 @@ func TestBlipSubChangesDocIDFilter(t *testing.T) {
 	subChangesRequest.Properties["batch"] = "10" // default batch size is 200, lower this to 5 to make sure we get multiple batches
 	subChangesRequest.SetCompressed(false)
 
-	body := subChangesBody{DocIDs: docIDsExpected}
+	body := replicator.SubChangesBody{DocIDs: docIDsExpected}
 	bodyBytes, err := base.JSONMarshal(body)
 	assert.NoError(t, err, "Error marshalling subChanges body.")
 
@@ -979,7 +980,7 @@ function(doc, oldDoc) {
 			panic(fmt.Sprintf("Unexpected err: %v", err))
 		}
 		_, isRemoved := doc[db.BodyRemoved]
-		assert.False(t, isRemoved, fmt.Sprintf("Document %v shouldn't be removed", request.Properties[revMessageId]))
+		assert.False(t, isRemoved, fmt.Sprintf("Document %v shouldn't be removed", request.Properties[replicator.RevMessageId]))
 
 	}
 
@@ -2007,7 +2008,7 @@ func TestBlipDeltaSyncPull(t *testing.T) {
 	// Check EE is delta, and CE is full-body replication
 	if base.IsEnterpriseEdition() {
 		// Check the request was sent with the correct deltaSrc property
-		assert.Equal(t, "1-0335a345b6ffed05707ccc4cbc1b67f4", msg.Properties[revMessageDeltaSrc])
+		assert.Equal(t, "1-0335a345b6ffed05707ccc4cbc1b67f4", msg.Properties[replicator.RevMessageDeltaSrc])
 		// Check the request body was the actual delta
 		msgBody, err := msg.Body()
 		assert.NoError(t, err)
@@ -2015,7 +2016,7 @@ func TestBlipDeltaSyncPull(t *testing.T) {
 		assert.Equal(t, deltaSentCount+1, base.ExpvarVar2Int(rt.GetDatabase().DbStats.StatsDeltaSync().Get(base.StatKeyDeltasSent)))
 	} else {
 		// Check the request was NOT sent with a deltaSrc property
-		assert.Equal(t, "", msg.Properties[revMessageDeltaSrc])
+		assert.Equal(t, "", msg.Properties[replicator.RevMessageDeltaSrc])
 		// Check the request body was NOT the delta
 		msgBody, err := msg.Body()
 		assert.NoError(t, err)
@@ -2127,7 +2128,7 @@ func TestBlipDeltaSyncPullTombstoned(t *testing.T) {
 	msgBody, err := msg.Body()
 	assert.NoError(t, err)
 	assert.Equal(t, `{}`, string(msgBody))
-	assert.Equal(t, "1", msg.Properties[revMessageDeleted])
+	assert.Equal(t, "1", msg.Properties[replicator.RevMessageDeleted])
 
 	deltaCacheHitsEnd := base.ExpvarVar2Int(rt.GetDatabase().DbStats.StatsDeltaSync().Get(base.StatKeyDeltaCacheHits))
 	deltaCacheMissesEnd := base.ExpvarVar2Int(rt.GetDatabase().DbStats.StatsDeltaSync().Get(base.StatKeyDeltaCacheMisses))
@@ -2220,13 +2221,13 @@ func TestBlipDeltaSyncPullTombstonedStarChan(t *testing.T) {
 
 	msg, ok := client1.WaitForBlipRevMessage("doc1", "2-ed278cbc310c9abeea414da15d0b2cac") // docid, revid to get the message
 	assert.True(t, ok)
-	assert.Equal(t, messageRev, msg.Profile(), "unexpected profile for message %v in %v",
+	assert.Equal(t, replicator.MessageRev, msg.Profile(), "unexpected profile for message %v in %v",
 		msg.SerialNumber(), client1.pullReplication.GetMessages())
 	msgBody, err := msg.Body()
 	assert.NoError(t, err)
 	assert.Equal(t, `{}`, string(msgBody), "unexpected body for message %v in %v",
 		msg.SerialNumber(), client1.pullReplication.GetMessages())
-	assert.Equal(t, "1", msg.Properties[revMessageDeleted], "unexpected deleted property for message %v in %v",
+	assert.Equal(t, "1", msg.Properties[replicator.RevMessageDeleted], "unexpected deleted property for message %v in %v",
 		msg.SerialNumber(), client1.pullReplication.GetMessages())
 
 	// Sync Gateway will have cached the tombstone delta, so client 2 should be able to retrieve it from the cache
@@ -2239,13 +2240,13 @@ func TestBlipDeltaSyncPullTombstonedStarChan(t *testing.T) {
 
 	msg, ok = client2.WaitForBlipRevMessage("doc1", "2-ed278cbc310c9abeea414da15d0b2cac")
 	assert.True(t, ok)
-	assert.Equal(t, messageRev, msg.Profile(), "unexpected profile for message %v in %v",
+	assert.Equal(t, replicator.MessageRev, msg.Profile(), "unexpected profile for message %v in %v",
 		msg.SerialNumber(), client2.pullReplication.GetMessages())
 	msgBody, err = msg.Body()
 	assert.NoError(t, err)
 	assert.Equal(t, `{}`, string(msgBody), "unexpected body for message %v in %v",
 		msg.SerialNumber(), client2.pullReplication.GetMessages())
-	assert.Equal(t, "1", msg.Properties[revMessageDeleted], "unexpected deleted property for message %v in %v",
+	assert.Equal(t, "1", msg.Properties[replicator.RevMessageDeleted], "unexpected deleted property for message %v in %v",
 		msg.SerialNumber(), client2.pullReplication.GetMessages())
 
 	deltaCacheHitsEnd := base.ExpvarVar2Int(rt.GetDatabase().DbStats.StatsDeltaSync().Get(base.StatKeyDeltaCacheHits))
@@ -2302,7 +2303,7 @@ func TestBlipPullRevMessageHistory(t *testing.T) {
 
 	msg, ok := client.pullReplication.WaitForMessage(5)
 	assert.True(t, ok)
-	assert.Equal(t, "1-0335a345b6ffed05707ccc4cbc1b67f4", msg.Properties[revMessageHistory])
+	assert.Equal(t, "1-0335a345b6ffed05707ccc4cbc1b67f4", msg.Properties[replicator.RevMessageHistory])
 }
 
 // TestBlipDeltaSyncPullRevCache tests that a simple pull replication uses deltas in EE,
@@ -2362,7 +2363,7 @@ func TestBlipDeltaSyncPullRevCache(t *testing.T) {
 
 	// Check EE is delta
 	// Check the request was sent with the correct deltaSrc property
-	assert.Equal(t, "1-0335a345b6ffed05707ccc4cbc1b67f4", msg.Properties[revMessageDeltaSrc])
+	assert.Equal(t, "1-0335a345b6ffed05707ccc4cbc1b67f4", msg.Properties[replicator.RevMessageDeltaSrc])
 	// Check the request body was the actual delta
 	msgBody, err := msg.Body()
 	assert.NoError(t, err)
@@ -2380,7 +2381,7 @@ func TestBlipDeltaSyncPullRevCache(t *testing.T) {
 	assert.True(t, ok)
 
 	// Check the request was sent with the correct deltaSrc property
-	assert.Equal(t, "1-0335a345b6ffed05707ccc4cbc1b67f4", msg2.Properties[revMessageDeltaSrc])
+	assert.Equal(t, "1-0335a345b6ffed05707ccc4cbc1b67f4", msg2.Properties[replicator.RevMessageDeltaSrc])
 	// Check the request body was the actual delta
 	msgBody2, err := msg2.Body()
 	assert.NoError(t, err)
@@ -2431,7 +2432,7 @@ func TestBlipDeltaSyncPush(t *testing.T) {
 
 	if base.IsEnterpriseEdition() {
 		// Check the request was sent with the correct deltaSrc property
-		assert.Equal(t, "1-0335a345b6ffed05707ccc4cbc1b67f4", msg.Properties[revMessageDeltaSrc])
+		assert.Equal(t, "1-0335a345b6ffed05707ccc4cbc1b67f4", msg.Properties[replicator.RevMessageDeltaSrc])
 		// Check the request body was the actual delta
 		msgBody, err := msg.Body()
 		assert.NoError(t, err)
@@ -2443,7 +2444,7 @@ func TestBlipDeltaSyncPush(t *testing.T) {
 		assert.NotContains(t, docRev.BodyBytes, "bob")
 	} else {
 		// Check the request was NOT sent with a deltaSrc property
-		assert.Equal(t, "", msg.Properties[revMessageDeltaSrc])
+		assert.Equal(t, "", msg.Properties[replicator.RevMessageDeltaSrc])
 		// Check the request body was NOT the delta
 		msgBody, err := msg.Body()
 		assert.NoError(t, err)
@@ -2526,7 +2527,7 @@ func TestBlipNonDeltaSyncPush(t *testing.T) {
 	assert.True(t, ok)
 
 	// Check the request was NOT sent with a deltaSrc property
-	assert.Equal(t, "", msg.Properties[revMessageDeltaSrc])
+	assert.Equal(t, "", msg.Properties[replicator.RevMessageDeltaSrc])
 	// Check the request body was NOT the delta
 	msgBody, err := msg.Body()
 	assert.NoError(t, err)
@@ -2595,14 +2596,14 @@ func TestBlipDeltaSyncNewAttachmentPull(t *testing.T) {
 
 	if base.IsEnterpriseEdition() {
 		// Check the request was sent with the correct deltaSrc property
-		assert.Equal(t, "1-0335a345b6ffed05707ccc4cbc1b67f4", msg.Properties[revMessageDeltaSrc])
+		assert.Equal(t, "1-0335a345b6ffed05707ccc4cbc1b67f4", msg.Properties[replicator.RevMessageDeltaSrc])
 		// Check the request body was the actual delta
 		msgBody, err := msg.Body()
 		assert.NoError(t, err)
 		assert.Equal(t, `{"_attachments":[{"hello.txt":{"digest":"sha1-Kq5sNclPz7QV2+lfQIuc6R7oRu0=","length":11,"revpos":2,"stub":true}}]}`, string(msgBody))
 	} else {
 		// Check the request was NOT sent with a deltaSrc property
-		assert.Equal(t, "", msg.Properties[revMessageDeltaSrc])
+		assert.Equal(t, "", msg.Properties[replicator.RevMessageDeltaSrc])
 		// Check the request body was NOT the delta
 		msgBody, err := msg.Body()
 		assert.NoError(t, err)
@@ -2679,25 +2680,25 @@ func TestBlipNorev(t *testing.T) {
 	require.NoError(t, err)
 	defer btc.Close()
 
-	norevMsg := NewNoRevMessage()
-	norevMsg.setId("docid")
-	norevMsg.setRev("1-a")
-	norevMsg.setError("404")
-	norevMsg.setReason("couldn't send xyz")
+	norevMsg := replicator.NewNoRevMessage()
+	norevMsg.SetId("docid")
+	norevMsg.SetRev("1-a")
+	norevMsg.SetError("404")
+	norevMsg.SetReason("couldn't send xyz")
 
 	// Couchbase Lite always sends noreply=true for norev messages
 	// but set to false so we can block waiting for a reply
 	norevMsg.SetNoReply(false)
 
 	// Request that the handler used to process the message is sent back in the response
-	norevMsg.Properties[sgShowHandler] = "true"
+	norevMsg.Properties[replicator.SGShowHandler] = "true"
 
 	assert.NoError(t, btc.pushReplication.sendMsg(norevMsg.Message))
 
 	// Check that the response we got back was processed by the norev handler
 	resp := norevMsg.Response()
 	assert.NotNil(t, resp)
-	assert.Equal(t, "handleNoRev", resp.Properties[sgHandler])
+	assert.Equal(t, "handleNoRev", resp.Properties[replicator.SGHandler])
 }
 
 // TestBlipDeltaSyncPushAttachment tests updating a doc that has an attachment with a delta that doesn't modify the attachment.
